fix: disable aws retry strategy
In the V2 AWS apis, in that AWS has a retry strategy if the request fails with something like a 503.
So by default the request will retry 3 times with a backoff wait period between each retry.
I think it could skew the results of a performance test.
AWS provides a noop retry implementation if you don't want retries at all.

Signed-off-by: Liang Zheng <zhengliang0901@gmail.com>